+++ src/shikafx-ext/__ccode__/utils/quicksort.h
@@ -7,12 +7,12 @@ extern "C" {
 
 #include "../common.h"
 
-// using int64_t, because skx_quicksort_typed_nums decrement low value at 0
+// using size_t, because skx_quicksort_typed_nums decrement low value at 0
 
-void skx_quicksort_typed_nums__swap(typed_nums_t* data, int64_t left, int64_t right);
-int64_t skx_quicksort_typed_nums__partition(typed_nums_t* data, int64_t low, int64_t high);
-void skx_quicksort_typed_nums__init(typed_nums_t* data, int64_t low, int64_t high);
-void skx_quicksort_typed_nums(typed_nums_t* data, int64_t size);
+void skx_quicksort_typed_nums__swap(typed_nums_t* data, size_t left, size_t right);
+size_t skx_quicksort_typed_nums__partition(typed_nums_t* data, size_t low, size_t high);
+void skx_quicksort_typed_nums__init(typed_nums_t* data, size_t low, size_t high);
+void skx_quicksort_typed_nums(typed_nums_t* data, size_t size);
 
 #ifdef __cplusplus
 }
